remove kotlin-dsl
