+++ build-logic/src/main/kotlin/main.kt
@@ -1,8 +1,12 @@
 import com.android.build.api.dsl.CommonExtension
+import com.android.build.gradle.internal.tasks.factory.dependsOn
 import dev.icerock.gradle.MultiplatformResourcesPluginExtension
-import io.openfeedback.OpenFeedback
+import internal.configurePublishingInternal
+import internal.publishIfNeededTaskProvider
+import internal.registerReleaseTask
 import org.gradle.api.JavaVersion
 import org.gradle.api.Project
+import org.gradle.api.logging.LogLevel
 import org.jetbrains.kotlin.gradle.dsl.KotlinCompile
 import org.jetbrains.kotlin.gradle.dsl.KotlinJvmOptions
 import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
@@ -26,12 +30,12 @@ private fun Project.configureAndroid(namespace: String) {
 
 private fun Project.configureKotlin() {
     tasks.withType(KotlinCompile::class.java) {
-        kotlinOptions {
+        it.kotlinOptions {
             if (this is KotlinJvmOptions) {
                 jvmTarget = JavaVersion.VERSION_17.toString()
             }
         }
-        this.kotlinOptions.freeCompilerArgs += "-Xexpect-actual-classes"
+        it.kotlinOptions.freeCompilerArgs += "-Xexpect-actual-classes"
     }
 }
 
@@ -59,9 +63,7 @@ fun Project.library(
     kotlin: KotlinMultiplatformExtension.() -> Unit = {}
 ) {
     if (artifactName != null) {
-        extensions.getByType(OpenFeedback::class.java).apply {
-            configurePublishing(artifactName)
-        }
+        configurePublishingInternal(artifactName)
     }
     configureAndroid(namespace = namespace)
     configureKMP()
@@ -79,4 +81,27 @@ fun Project.androidApp(
 ) {
     configureAndroid(namespace = namespace)
     configureKotlin()
+}
+
+
+fun Project.configureRoot() {
+    check(this == rootProject) {
+        "configureRoot must be called from the root project"
+    }
+
+    val publishIfNeeded = project.publishIfNeededTaskProvider()
+    val ossStagingReleaseTask = project.registerReleaseTask("ossStagingRelease")
+
+    val eventName = System.getenv(EnvVarKeys.GitHub.event)
+    val ref = System.getenv(EnvVarKeys.GitHub.ref)
+
+    if (eventName == "push" && ref == "refs/heads/main" && project.version.toString().endsWith("SNAPSHOT")) {
+        project.logger.log(LogLevel.LIFECYCLE, "Deploying snapshot to OssSnapshot...")
+        publishIfNeeded.dependsOn(project.tasks.named("publishAllPublicationsToOssSnapshotsRepository"))
+    }
+
+    if (ref?.startsWith("refs/tags/") == true) {
+        project.logger.log(LogLevel.LIFECYCLE, "Deploying release to OssStaging...")
+        publishIfNeeded.dependsOn(ossStagingReleaseTask)
+    }
 }
